pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "uecs-llm"
version = "0.1.0"
description = "LLM-based greenhouse environment control with UECS-CCM integration"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    { name = "Yasunori Oi" },
]
keywords = ["uecs", "llm", "greenhouse", "agriculture", "iot", "unipi"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "pyyaml>=6.0",
    "httpx>=0.24.0",
]

[project.optional-dependencies]
daemon = [
    "paho-mqtt>=1.6,<3",
    "smbus2>=0.4",
    "gpiod>=2.0",
    "pyserial>=3.5",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "httpx>=0.24.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
]

[project.scripts]
unipi-daemon = "unipi_daemon.main:main"
agriha-control = "uecs_llm.agriha_control:main"

[tool.hatch.build.targets.wheel]
packages = ["src/uecs_llm", "src/unipi_daemon"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
